review changes
- add a new `LocalizedTexts`
- remove read restext as dictionary extension
- remove inline fill locales method
