+++ src/recipe/Baked.Recipe.Service.Application/Localization/AspNetCore/AspNetCoreLocalizationExtensions.cs
@@ -15,25 +15,4 @@ public static class AspNetCoreLocalizationExtensions
 
         return new(language, otherLanguages);
     }
-
-    public static Dictionary<string, string> ReadResxFileAsDictionary(string path)
-    {
-        var values = new Dictionary<string, string>();
-        using (StreamReader reader = new(path))
-        {
-            string? line = reader.ReadLine();
-            while (line != null)
-            {
-                if (!string.IsNullOrWhiteSpace(line.Trim()))
-                {
-                    var keyValue = line.Split('=', StringSplitOptions.TrimEntries);
-                    values.TryAdd(keyValue[0], keyValue[1]);
-                }
-
-                line = reader.ReadLine();
-            }
-        }
-
-        return values;
-    }
 }
